refactor(sdk-coin-apt): transaction builder
TICKET: COIN-2896
