+++ modules/sdk-coin-apt/src/lib/constants.ts
@@ -5,3 +5,6 @@ export const APT_SIGNATURE_LENGTH = 128;
 export const UNAVAILABLE_TEXT = 'UNAVAILABLE';
 export const DEFAULT_GAS_UNIT_PRICE = 100;
 export const SECONDS_PER_WEEK = 7 * 24 * 60 * 60; // Days * Hours * Minutes * Seconds
+
+export const APTOS_ACCOUNT_MODULE = 'aptos_account';
+export const FUNGIBLE_ASSET_MODULE = 'fungible_asset';
