+++ modules/sdk-coin-apt/src/lib/transferBuilder.ts
@@ -2,6 +2,7 @@ import { TransactionBuilder } from './transactionBuilder';
 import { BaseCoin as CoinConfig } from '@bitgo/statics';
 import { TransactionType } from '@bitgo/sdk-core';
 import { TransferTransaction } from './transaction/transferTransaction';
+import { Transaction } from './transaction/transaction';
 
 export class TransferBuilder extends TransactionBuilder {
   constructor(_coinConfig: Readonly<CoinConfig>) {
@@ -20,5 +21,19 @@ export class TransferBuilder extends TransactionBuilder {
    */
   initBuilder(tx: TransferTransaction): void {
     this._transaction = tx;
+  }
+
+  /** @inheritdoc */
+  protected fromImplementation(rawTransaction: string): Transaction {
+    this.transaction.fromRawTransaction(rawTransaction);
+    this.transaction.transactionType = this.transactionType;
+    return this.transaction;
+  }
+
+  /** @inheritdoc */
+  protected async buildImplementation(): Promise<Transaction> {
+    this.transaction.transactionType = this.transactionType;
+    await this.transaction.build();
+    return this.transaction;
   }
 }
